pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "phoenix-trade-bot"
version = "0.5.0"
description = "Enterprise phoenix trade bot platform"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "aiokafka>=0.10.0",
    "msgpack>=1.0.0",
    "redis[hiredis]>=5.0.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "email-validator>=2.0.0",
    "pydantic-settings>=2.1.0",
    "bcrypt>=4.1.0",
    "python-jose[cryptography]>=3.3.0",
    "cryptography>=42.0.0",
    "alpaca-py>=0.21.0",
    "discord.py-self>=2.1.0",
    "prometheus-client>=0.20.0",
    "pyyaml>=6.0.0",
    "pyotp>=2.9.0",
    "jinja2>=3.1.0",
    "claude-agent-sdk>=0.1.50",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
    "pyarrow>=15.0.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "catboost>=1.2.0",
    "joblib>=1.3.0",
    "torch>=2.2.0",
    "sentence-transformers>=3.0.0",
    "anthropic>=0.40.0",
    "apscheduler>=3.10.0",
    "croniter>=2.0.0",
    "gunicorn>=22.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "pytest-playwright>=0.4.0",
    "respx>=0.20.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "httpx>=0.27.0",
    "aiosqlite>=0.19.0",
    "rapidfuzz>=3.0.0",
    "numpy>=1.26.0",
    "pandas>=2.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "apps/api/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
markers = ["benchmark: benchmark test"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "services", "shared"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["sqlalchemy.ext.mypy.plugin"]

[tool.setuptools.packages.find]
include = ["shared*"]
